added util tests
